Refactored sites module for better naming consistency
