+++ editor/src/app/sites/settings/site-settings.module.ts
@@ -1,18 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SiteSettingsComponent } from './site-settings.component';
-import { SitesSettingsState } from './sites-settings.state';
+import { SiteSettingsState } from './site-settings.state';
 import { NgxsModule } from '@ngxs/store';
 
 @NgModule({
   imports: [
     CommonModule,
     NgxsModule.forFeature([
-      SitesSettingsState
+      SiteSettingsState
     ])
   ],
   declarations: [
     SiteSettingsComponent
   ]
 })
-export class SitesSettingsModule { }
+export class SiteSettingsModule { }
